backend/requirements.txt
# Web framework + server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Auth
python-jose[cryptography]>=3.3.0

# HTTP client (Google OAuth server-to-server exchange)
httpx>=0.25.0

# Redis (one-time code store)
redis>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Validation
pydantic[email]>=2.7.0

# Dev / test / lint (not needed on Render)
black>=24.0.0
ruff>=0.4.0
mypy>=1.9.0
pytest>=8.0.0
factory_boy>=3.3.0
